requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
openai>=1.0.0
numpy>=1.24.0
pyttsx3>=2.90
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.13 